Merged PR 8466801: Creation of Voting Procedure Command
In this PR, I've created a basic framework for the voting procedure of a proposal after a member/user ID certificates and proposal json file have been generated. Please feel free to leave any feedback on how to optimize the process, specifically the process of inputting the proposal ID and the selection choices of the proposal files.

Related work items: #24602734