Merged PR 8498742: List All Active Proposals
In this PR, I have added onto the votingProcedure.ts command where instead of entering the proposal ID into an input box, the user will be shown all of the active proposals that haven't been voted on via dropdown quick pick. From there, the user will click on the proposal ID they'd like to cast a vote for and the ID string will be returned and passed through the script. Has been tested on live running MCCF instance scondelli-721

The function for this is under **displayProposals** function.

In the generate identity and osUtilities files I made small fixes to the code.

Related work items: #24640354, #24663046