Merged PR 8513868: Setup basic build pipelines for extension code
Setup build pipelines (PR and Official) compliant with OneBranch to install, build, and package the VS Code extension.

The build stage will run for both Linux and Windows platforms. Step to run tests has been added but it's currently excluded due to some undebuggable failures in the pipeline. Nonetheless, we do not have any relevant tests at the moment, so it safe to exclude at this point.

The package step will publish an artifacts containing the VSIX file for the extension.

[Example build pipeline](https://msazure.visualstudio.com/One/_build/results?buildId=77245515&view=results)

![image.png](https://msazure.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/57bfbddf-4c2c-46ec-a6b1-6049605c642d/pullRequests/8513868/attachments/image.png)

[Artifacts produced](https://msazure.visualstudio.com/One/_build/results?buildId=77245515&view=artifacts&pathAsName=false&type=publishedArtifacts)

![image (2).png](https://msazure.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/57bfbddf-4c2c-46ec-a6b1-6049605c642d/pullRequests/8513868/attachments/image%20%282%29.png)

Related work items: #24674060