Merged PR 9050697: A command to test operator actions in custom constitution
There are certain operator member actions that are essential to manage a Managed CCF network. Customers that want to deploy custom constitution to their Managed CCF resource should ensure that the operator actions succeed in the local deployment before deploying it to the cloud resource.

This PR contains an action to run a few tests on a local CCF network with the custom constitution deployed on it. It will print the result of the tests. If the tests passed, it confirms that an operator member in the Managed CCF resource can perform the actions.

# Testing/Validation

- [ ] Manually validated the changes by running the extension locally

## Scenarios validated

I tested the operator actions on a 1 node CCF network that was deployed with the 'Start CCF network in a devcontainer' command.

# Documentation

Yes

# Type of change

- [ ] Feature work

## Does this introduce a breaking change?

- [ ] No

## Risk

- [ ] Small – No issues are expected.

Related work items: #24715412